+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version (NEXT)
+* use a single state atom in simulator
 
 ## Version 0.1.0
 * implemented quantom states, quantum gates, gate registry and quantum cirquits